modified:   client/main.py modified:   server/learning_log_server.py Added ability to override home directory detection by providing a new path in the client_settings..txt file for client.
Server now merges new log entries instead of simply overwriting each time.

Added class handling via a json file. This is a dictionary of classes with a list of pupils in each class. This must be modified by hand for now, but a tool to make this less painfull will be made as time permits.